Task2_LongHairIdentification/requirements..txt
tensorflow==2.5.0
scikit-learn==0.24.2
numpy==1.19.5
Pillow==8.3.1
opencv-python==4.5.3.56